+++ src/test_e2e/customer_id/200_get_by_id.test.ts
@@ -1,5 +1,5 @@
 import request from "supertest";
-import { app, pool } from "../../utils/app";
+import { app, pool } from "../../app";
 afterEach(() => {
 	pool.end();
 });
